Move SelectInput.test into Select.test
